+++ src/main/java/com/github/jlangch/venice/impl/functions/IOFnBlacklisted.java
@@ -36,6 +36,7 @@ public class IOFnBlacklisted {
 
 								// system
 								"gc",
+								"shutdown-hook",
 								"sleep",
 								"sh",
 								"system-prop",