+++ ChangeLog.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 
 - queue datatype (bounded or unbounded)
+- Venice function to be used as JVM shutdown hook
 
 ### Fixed